Use Double everywhere

+++ src/ext/java/org/opentripplanner/ext/legacygraphqlapi/datafetchers/LegacyGraphQLQueryTypeImpl.java
@@ -722,7 +722,7 @@ public class LegacyGraphQLQueryTypeImpl
       callWith.argument("requiredVehicleParkingTags", (Collection<String> v) -> request.requiredVehicleParkingTags = new HashSet<>(v));
       callWith.argument("bannedVehicleParkingTags", (Collection<String> v) -> request.bannedVehicleParkingTags = new HashSet<>(v));
       callWith.argument("preferredVehicleParkingTags", (Collection<String> v) -> request.preferredVehicleParkingTags = new HashSet<>(v));
-      callWith.argument("unpreferredVehicleParkingTagPenalty", (Float v) -> request.unpreferredVehicleParkingTagPenalty = v);
+      callWith.argument("unpreferredVehicleParkingTagPenalty", (Double v) -> request.unpreferredVehicleParkingTagPenalty = v);
 
       callWith.argument(
               "modeWeight", (Map<String, Object> v) -> request.setTransitReluctanceForMode(
